Use the shared validation tools moved to knative/pkg.
